# tests/GenAIPlatform.IntegrationTests/Baselines/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GenAIPlatform.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/chat/direct": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Run a direct (non-RAG) chat completion against the configured model gateway.",
        "operationId": "CreateDirectChatCompletion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectChatHttpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/rag": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Run a retrieval-augmented chat completion using indexed document context.",
        "operationId": "CreateRagChatCompletion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RagChatHttpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RagChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/agentic": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Run an agentic chat completion: the model may propose backend tool calls subject to policy.",
        "operationId": "CreateAgenticChatCompletion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgenticChatHttpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgenticChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Upload a document for ingestion (indexed asynchronously by the worker).",
        "operationId": "UploadDocument",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/IFormFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDocumentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{documentId}": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Get the indexing status of a previously uploaded document.",
        "operationId": "GetDocumentStatus",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgenticChatHttpRequest": {
        "required": [
          "message",
          "model",
          "temperature",
          "maxOutputTokens",
          "correlationId",
          "approveRiskyTools"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "model": {
            "type": [
              "null",
              "string"
            ]
          },
          "temperature": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "maxOutputTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "approveRiskyTools": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
      "AgenticChatResponse": {
        "required": [
          "conversationId",
          "status",
          "answer",
          "model",
          "provider",
          "correlationId",
          "steps",
          "toolCalls",
          "totalTokens",
          "estimatedCost",
          "toolResults",
          "usage"
        ],
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "answer": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "provider": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": "string"
          },
          "steps": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "toolCalls": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "estimatedCost": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "toolResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentToolCallResult"
            }
          },
          "usage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AiModelUsage"
              }
            ]
          }
        }
      },
      "AgentToolCallResult": {
        "required": [
          "toolCallId",
          "toolName",
          "schemaVersion",
          "policyDecision",
          "executionStatus",
          "result",
          "errorCode"
        ],
        "type": "object",
        "properties": {
          "toolCallId": {
            "type": "string"
          },
          "toolName": {
            "type": "string"
          },
          "schemaVersion": {
            "type": "string"
          },
          "policyDecision": {
            "type": "string"
          },
          "executionStatus": {
            "$ref": "#/components/schemas/ToolExecutionStatus"
          },
          "result": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorCode": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "AiModelUsage": {
        "required": [
          "inputTokens",
          "outputTokens",
          "totalTokens"
        ],
        "type": "object",
        "properties": {
          "inputTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "outputTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "DirectChatHttpRequest": {
        "required": [
          "message",
          "model",
          "temperature",
          "maxOutputTokens",
          "correlationId"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "model": {
            "type": [
              "null",
              "string"
            ]
          },
          "temperature": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "maxOutputTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "DirectChatResponse": {
        "required": [
          "message",
          "model",
          "provider",
          "usage",
          "prompt",
          "correlationId"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "usage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AiModelUsage"
              }
            ]
          },
          "prompt": {
            "$ref": "#/components/schemas/PromptMetadata"
          },
          "correlationId": {
            "type": "string"
          }
        }
      },
      "DocumentStatusResponse": {
        "required": [
          "documentId",
          "title",
          "fileName",
          "version",
          "accessLevel",
          "indexingStatus",
          "indexingJobId",
          "indexingJobStatus",
          "indexingAttempts",
          "chunkCount",
          "failureReason",
          "updatedAtUtc"
        ],
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "accessLevel": {
            "type": "string"
          },
          "indexingStatus": {
            "type": "string"
          },
          "indexingJobId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "indexingJobStatus": {
            "type": [
              "null",
              "string"
            ]
          },
          "indexingAttempts": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "chunkCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "failureReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
      "JsonElement": { },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PromptMetadata": {
        "required": [
          "templateName",
          "version",
          "contentHash"
        ],
        "type": "object",
        "properties": {
          "templateName": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "contentHash": {
            "type": "string"
          }
        }
      },
      "RagChatHttpRequest": {
        "required": [
          "message",
          "model",
          "temperature",
          "maxOutputTokens",
          "topK",
          "minSimilarityScore",
          "documentIds",
          "correlationId"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "model": {
            "type": [
              "null",
              "string"
            ]
          },
          "temperature": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "maxOutputTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "topK": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "minSimilarityScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "documentIds": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RagChatResponse": {
        "required": [
          "message",
          "model",
          "provider",
          "usage",
          "prompt",
          "correlationId",
          "noContext",
          "citations"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "provider": {
            "type": [
              "null",
              "string"
            ]
          },
          "usage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AiModelUsage"
              }
            ]
          },
          "prompt": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PromptMetadata"
              }
            ]
          },
          "correlationId": {
            "type": "string"
          },
          "noContext": {
            "type": "boolean"
          },
          "citations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RagCitation"
            }
          }
        }
      },
      "RagCitation": {
        "required": [
          "referenceId",
          "documentId",
          "chunkId",
          "documentVersion",
          "chunkPosition",
          "title",
          "fileName",
          "similarityScore"
        ],
        "type": "object",
        "properties": {
          "referenceId": {
            "type": "string"
          },
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "chunkId": {
            "type": "string",
            "format": "uuid"
          },
          "documentVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "chunkPosition": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "similarityScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "ToolExecutionStatus": {
        "enum": [
          "NotExecuted",
          "Rejected",
          "ValidationFailed",
          "ApprovalRequired",
          "Failed",
          "Succeeded"
        ]
      },
      "UploadDocumentResponse": {
        "required": [
          "documentId",
          "title",
          "fileName",
          "version",
          "accessLevel",
          "indexingStatus",
          "indexingJobId",
          "indexingJobStatus",
          "createdAtUtc"
        ],
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "accessLevel": {
            "type": "string"
          },
          "indexingStatus": {
            "type": "string"
          },
          "indexingJobId": {
            "type": "string",
            "format": "uuid"
          },
          "indexingJobStatus": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}